push 8/7

+++ Assets/Scripts/MercanciaScript.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class MercanciaScript : MonoBehaviour
 {
-    public int precio;
+    public float precio;
     public PlayerMoney dineroManager;
 
     void Start()
